#1569 Change to improve the new `Data Delete` task
Additional changes to improve the new `Data Delete` task that replaces
the `File System Clean` task.

+++ stroom-data/stroom-data-store-impl-fs/src/main/java/stroom/data/store/impl/fs/PhysicalDeleteExecutor.java
@@ -26,8 +26,11 @@ import stroom.meta.shared.MetaFields;
 import stroom.meta.shared.Status;
 import stroom.query.api.v2.ExpressionOperator;
 import stroom.query.api.v2.ExpressionTerm.Condition;
+import stroom.task.api.ExecutorProvider;
 import stroom.task.api.TaskContext;
 import stroom.task.api.TaskContextFactory;
+import stroom.task.api.ThreadPoolImpl;
+import stroom.task.shared.ThreadPool;
 import stroom.util.date.DateUtil;
 import stroom.util.io.FileUtil;
 import stroom.util.logging.LambdaLogUtil;
@@ -42,9 +45,18 @@ import java.io.UncheckedIOException;
 import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Queue;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.Executor;
+import java.util.concurrent.LinkedBlockingQueue;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 
 public class PhysicalDeleteExecutor {
     private static final LambdaLogger LOGGER = LambdaLoggerFactory.getLogger(PhysicalDeleteExecutor.class);
@@ -59,6 +71,8 @@ public class PhysicalDeleteExecutor {
     private final PhysicalDelete physicalDelete;
     private final DataVolumeDao dataVolumeDao;
     private final TaskContextFactory taskContextFactory;
+    private final ExecutorProvider executorProvider;
+    private final DataStoreServiceConfig config;
 
     @Inject
     PhysicalDeleteExecutor(
@@ -68,7 +82,9 @@ public class PhysicalDeleteExecutor {
             final MetaService metaService,
             final PhysicalDelete physicalDelete,
             final DataVolumeDao dataVolumeDao,
-            final TaskContextFactory taskContextFactory) {
+            final TaskContextFactory taskContextFactory,
+            final ExecutorProvider executorProvider,
+            final DataStoreServiceConfig config) {
         this.clusterLockService = clusterLockService;
         this.dataStoreServiceConfig = dataStoreServiceConfig;
         this.fileSystemStreamPathHelper = fileSystemStreamPathHelper;
@@ -76,6 +92,8 @@ public class PhysicalDeleteExecutor {
         this.physicalDelete = physicalDelete;
         this.dataVolumeDao = dataVolumeDao;
         this.taskContextFactory = taskContextFactory;
+        this.executorProvider = executorProvider;
+        this.config = config;
     }
 
     public void exec() {
@@ -102,24 +120,42 @@ public class PhysicalDeleteExecutor {
 
     public void delete(final TaskContext taskContext, final long deleteThresholdEpochMs) {
         if (!Thread.currentThread().isInterrupted()) {
-            long count = 0;
+            long count;
             long total = 0;
 
             final LogExecutionTime logExecutionTime = new LogExecutionTime();
 
             final int deleteBatchSize = dataStoreServiceConfig.getDeleteBatchSize();
 
+            long minId = 0;
             if (!Thread.currentThread().isInterrupted()) {
+                final ThreadPool threadPool = new ThreadPoolImpl("Data Delete#", 1, 1, config.getFileSystemCleanBatchSize(), Integer.MAX_VALUE);
+                final Executor executor = executorProvider.get(threadPool);
+
                 do {
-                    // Insert a batch of ids into the temp id table and find out
-                    // how many were inserted.
-                    final List<Meta> idList = getDeleteIdList(deleteThresholdEpochMs, deleteBatchSize);
-                    count = idList.size();
+                    // Increment the minimum id.
+                    //
+                    // Using a minimum id ensures we don't get stuck if we are unable to delete some meta data due to
+                    // difficulties removing files.
+                    minId++;
+
+                    // Get a batch of meta ids that are ready for actual deletion.
+                    final List<Meta> metaList = getDeleteList(minId, deleteThresholdEpochMs, deleteBatchSize);
+                    count = metaList.size();
 
                     // If we inserted some ids then try and delete this batch.
                     if (count > 0) {
+                        // Calculate the next minimum id to query for.
+                        final Optional<Long> maxId = metaList
+                                .stream()
+                                .map(Meta::getId)
+                                .max(Comparator.naturalOrder());
+                        if (maxId.isPresent()) {
+                            minId = Math.max(minId, maxId.get());
+                        }
+
                         total += count;
-                        deleteCurrentBatch(taskContext, idList);
+                        deleteCurrentBatch(taskContext, metaList, deleteThresholdEpochMs, executor);
                     }
                 } while (!Thread.currentThread().isInterrupted() && count >= deleteBatchSize);
             }
@@ -128,48 +164,37 @@ public class PhysicalDeleteExecutor {
         }
     }
 
-    private void deleteCurrentBatch(final TaskContext taskContext, final List<Meta> metaList) {
+    private void deleteCurrentBatch(final TaskContext taskContext, final List<Meta> metaList, final long deleteThresholdEpochMs, final Executor executor) {
         try {
+            final LinkedBlockingQueue<Long> successfulMetaIdDeleteQueue = new LinkedBlockingQueue<>();
+            final Map<Path, Path> directoryMap = new ConcurrentHashMap<>();
+
             // Delete all matching files.
+            final List<CompletableFuture<Void>> futures = new ArrayList<>();
             for (final Meta meta : metaList) {
-                info(taskContext, () -> "Deleting everything associated with " + meta);
-
-                final DataVolume dataVolume = dataVolumeDao.findDataVolume(meta.getId());
-                if (dataVolume == null) {
-                    LOGGER.warn(() -> "Unable to find any volume for " + meta);
-                } else {
-                    final Path file = fileSystemStreamPathHelper.getRootPath(dataVolume.getVolumePath(), meta, meta.getTypeName());
-                    final Path dir = file.getParent();
-                    String baseName = file.getFileName().toString();
-                    baseName = baseName.substring(0, baseName.indexOf("."));
-
-                    try (final DirectoryStream<Path> stream = Files.newDirectoryStream(dir, baseName + ".*")) {
-                        stream.forEach(f -> {
-                            try {
-                                info(taskContext, () -> "Deleting file: " + FileUtil.getCanonicalPath(f));
-                                Files.deleteIfExists(f);
-
-                            } catch (final InterruptedException e) {
-                                LOGGER.debug(e::getMessage, e);
-
-                                // Continue to interrupt.
-                                Thread.currentThread().interrupt();
-
-                                throw new RuntimeException(e);
-
-                            } catch (final IOException e) {
-                                throw new UncheckedIOException(e);
-                            }
-                        });
-                    } catch (final IOException e) {
-                        throw new UncheckedIOException(e);
-                    }
-                }
+                final Runnable runnable = deleteFiles(meta, taskContext, successfulMetaIdDeleteQueue, directoryMap);
+                final CompletableFuture<Void> completableFuture = CompletableFuture.runAsync(runnable, executor);
+                futures.add(completableFuture);
             }
 
-            final List<Long> metaIdList = metaList.stream().map(Meta::getId).collect(Collectors.toList());
+            // Wait for all completable futures to complete.
+            CompletableFuture.allOf(futures.toArray(new CompletableFuture[0])).join();
+            if (Thread.interrupted()) {
+                throw new InterruptedException();
+            }
 
-            // Delete meta volumes.
+            // Cleanup empty directories.
+            directoryMap.forEach((dir, root) ->
+                    tryDeleteDir(root, dir, deleteThresholdEpochMs));
+
+            if (Thread.interrupted()) {
+                throw new InterruptedException();
+            }
+
+            final List<Long> metaIdList = new ArrayList<>();
+            successfulMetaIdDeleteQueue.drainTo(metaIdList);
+
+            // Delete data volumes.
             info(taskContext, () -> "Deleting data volumes");
             dataVolumeDao.delete(metaIdList);
 
@@ -185,19 +210,98 @@ public class PhysicalDeleteExecutor {
         }
     }
 
-    private void info(final TaskContext taskContext, final Supplier<String> message) throws InterruptedException {
+    private void tryDeleteDir(final Path root,
+                              final Path dir,
+                              final long oldFileTime) {
+        try {
+            final String canonicalRoot = FileUtil.getCanonicalPath(root);
+            final String canonicalDir = FileUtil.getCanonicalPath(dir);
+
+            if (canonicalRoot.length() > 2
+                    && canonicalDir.startsWith(canonicalRoot)
+                    && !Files.isSameFile(root, dir)) {
+                final long lastModified = Files.getLastModifiedTime(dir).toMillis();
+
+                if (lastModified < oldFileTime) {
+                    try {
+                        Files.delete(dir);
+                        LOGGER.debug("tryDelete() - Deleted dir {}", canonicalDir);
+
+                        // Recurse.
+                        tryDeleteDir(root, dir.getParent(), oldFileTime);
+                    } catch (final IOException e) {
+                        LOGGER.debug("tryDelete() - Failed to delete dir {}", canonicalDir);
+                    }
+
+                } else {
+                    LOGGER.debug("tryDelete() - Dir too new to delete {}", canonicalDir);
+                }
+            }
+        } catch (final IOException e) {
+            LOGGER.error("tryDelete() - Failed to delete dir {}", FileUtil.getCanonicalPath(dir), e);
+        }
+    }
+
+    private Runnable deleteFiles(final Meta meta,
+                                 final TaskContext parentTaskContext,
+                                 final Queue<Long> successfulMetaIdDeleteQueue,
+                                 final Map<Path, Path> directoryMap) {
+        return taskContextFactory.context(parentTaskContext, "Deleting files", taskContext -> {
+            try {
+                if (Thread.interrupted()) {
+                    throw new InterruptedException();
+                }
+
+                info(taskContext, () -> "Deleting everything associated with " + meta);
+
+                final DataVolume dataVolume = dataVolumeDao.findDataVolume(meta.getId());
+                if (dataVolume == null) {
+                    LOGGER.warn(() -> "Unable to find any volume for " + meta);
+
+                } else {
+                    final Path volumePath = Paths.get(dataVolume.getVolumePath());
+                    final Path file = fileSystemStreamPathHelper.getRootPath(volumePath, meta, meta.getTypeName());
+                    final Path dir = file.getParent();
+                    String baseName = file.getFileName().toString();
+                    baseName = baseName.substring(0, baseName.indexOf("."));
+
+                    try (final DirectoryStream<Path> stream = Files.newDirectoryStream(dir, baseName + ".*")) {
+                        stream.forEach(f -> {
+                            try {
+                                info(taskContext, () -> "Deleting file: " + FileUtil.getCanonicalPath(f));
+                                Files.deleteIfExists(f);
+                            } catch (final IOException e) {
+                                throw new UncheckedIOException(e);
+                            }
+                        });
+                    } catch (final IOException e) {
+                        throw new UncheckedIOException(e);
+                    }
+
+                    directoryMap.put(dir, volumePath);
+                }
+
+                successfulMetaIdDeleteQueue.add(meta.getId());
+
+            } catch (final InterruptedException e) {
+                LOGGER.debug(e::getMessage, e);
+
+                // Continue to interrupt.
+                Thread.currentThread().interrupt();
+            }
+        });
+    }
+
+    private void info(final TaskContext taskContext, final Supplier<String> message) {
         try {
             taskContext.info(message);
             LOGGER.debug(message);
         } catch (final RuntimeException e) {
             LOGGER.error(e::getMessage, e);
         }
-        if (Thread.interrupted()) {
-            throw new InterruptedException();
-        }
     }
 
-    private List<Meta> getDeleteIdList(final long deleteThresholdEpochMs, final int batchSize) {
+    private List<Meta> getDeleteList(final long minId, final long deleteThresholdEpochMs, final int batchSize) {
         final ExpressionOperator expression = new ExpressionOperator.Builder()
                 .addTerm(
                         MetaFields.STATUS,
@@ -207,10 +311,15 @@ public class PhysicalDeleteExecutor {
                         MetaFields.STATUS_TIME,
                         Condition.LESS_THAN,
                         DateUtil.createNormalDateTimeString(deleteThresholdEpochMs))
+                .addTerm(
+                        MetaFields.ID,
+                        Condition.GREATER_THAN_OR_EQUAL_TO,
+                        minId
+                )
                 .build();
 
         final FindMetaCriteria criteria = new FindMetaCriteria(expression);
-        criteria.setSort(MetaFields.ID.getDisplayValue());
+        criteria.setSort(MetaFields.ID.getName());
         criteria.obtainPageRequest().setLength(batchSize);
 
         return metaService.find(criteria).getValues();
